requirements.txt
google-generativeai
gTTS
pyperclip
python-dotenv
google-auth-oauthlib
google-api-python-client
google-auth
google-cloud-texttospeech
